modifiers/base: rename from modifiers/index
